+++ src/main/java/ca/mcmaster/se2aa4/island/team43/Drone/Drone.java
@@ -1,4 +1,8 @@
-package ca.mcmaster.se2aa4.island.team43;
+package ca.mcmaster.se2aa4.island.team43.Drone;
+
+import ca.mcmaster.se2aa4.island.team43.Drone.*;
+import ca.mcmaster.se2aa4.island.team43.HomeBase.*;
+import ca.mcmaster.se2aa4.island.team43.Map.*;
 
 import java.util.Map;
 import java.util.HashMap;
