+++ src/main/java/ca/mcmaster/se2aa4/island/team43/Drone/Orientation.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.team43;
+package ca.mcmaster.se2aa4.island.team43.Drone;
 
 /**
  * The Orientation enum represents the four directions and methods when turning right or left.
