+++ src/main/java/ca/mcmaster/se2aa4/island/team43/HomeBase/JSONHandler.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.team43;
+package ca.mcmaster.se2aa4.island.team43.HomeBase;
 
 import org.json.*;
 
@@ -10,7 +10,7 @@ import org.json.JSONTokener;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import ca.mcmaster.se2aa4.island.team43.CommandCenter;
+import ca.mcmaster.se2aa4.island.team43.HomeBase.*;
 
 public class JSONHandler {
