+++ src/main/java/ca/mcmaster/se2aa4/island/team43/Map/Island.java
@@ -1,4 +1,6 @@
-package ca.mcmaster.se2aa4.island.team43;
+package ca.mcmaster.se2aa4.island.team43.Map;
+
+import ca.mcmaster.se2aa4.island.team43.Map.*;
 
 import java.util.ArrayList;
