+++ src/main/java/ca/mcmaster/se2aa4/island/team43/Map/POI.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.team43;
+package ca.mcmaster.se2aa4.island.team43.Map;
 
 public interface POI {
     //Points of interest
